requirements/base.txt
# Django and DRF
Django>=5.0.2
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
django-environ>=0.11.2

# Database
psycopg2-binary>=2.9.9

# Task Queue
celery>=5.3.6
redis>=5.0.1
django-celery-beat>=2.5.0
django-celery-results>=2.5.1

# Image Processing
Pillow>=10.2.0

# Payment Processing
stripe>=7.12.0

# API Documentation
drf-spectacular>=0.27.1

# Caching
django-redis>=5.4.0

# Security
django-axes>=6.3.0

# Add this line
drf-nested-routers>=0.93.4 